requirements.txt
torch
torchvision
streamlit
Pillow
numpy
baal